Move payloads to separate file
